pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "reolink_api"
version = '1.0.1'
description = "Unofficial Reolink API fork for personal use in Python 3.10"
authors = ["bobrock <bobrock@tuta.io>"]
license = "GPL-3.0"
readme = 'README.md'
repository = 'https://github.com/barretobrock/reolink-python-api'
packages = [
    { include = 'reolink_api' },
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "9.1.0"
PySocks = "1.7.1"
PyYAML = "6.0"
loguru = "0.6.0"
numpy = "^1.22.3"
opencv-python = "4.5.5.64"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
tox = "^3.24.5"

[tool.poetry.extras]
test = ["pytest"]
